+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [0.2.0]
+
+### Added
+
+- Support for asset processing
+- postcard format
+
+### Changed
+
+- Renamed `JsonReflectAssetPlugin` to `ReflectAssetPlugin`
+- Bump Rust edition to 2024
+
 ## [0.1.1]
 
 ### Fixed